requirements.txt
requests>=2.31.0
cryptography>=41.0.0
python-dotenv>=1.0.0
schedule>=1.2.0
numpy>=1.24.0
pydantic>=2.0.0
pytest>=7.4.0
pytest-mock>=3.12.0
ruff>=0.1.0
